feat: 移动文件位置, 调整为标准go项目结构
