--- go/changelog.md
@@ -1,1 +1,0 @@
-# v0.0.1
