--- go/main.go
@@ -1,8 +1,0 @@
-package main
-
-import "go2fe/test"
-
-func main() {
-	test.Run()
-	return
-}
