--- go/runner.conf
@@ -1,12 +1,0 @@
-root:              .
-tmp_path:          .tmp
-build_name:        app-go2fe
-build_log:         app-go2fe-build-errors.log
-valid_ext:         .go, .tmpl
-ignored:           cache, log, client, tmp, bin, static
-build_delay:       600
-colors:            1
-log_color_main:    cyan
-log_color_build:   yellow
-log_color_runner:  green
-log_color_watcher: magenta
