+++ main.go
@@ -1,0 +1,8 @@
+package main
+
+import "go2fe/test"
+
+func main() {
+	test.Run()
+	return
+}
